% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AdRevenue}
\alias{AdRevenue}
\title{U.S. Magazine Ad Revenue in 2006}
\format{A data frame with 70 observations on the following 5 variables:
\describe{
\item{Magazine}{Magazine name}
\item{ParentCompany}{Parent Company of the magazine}
\item{Subsidiary}{Subsidiary of the parent company, if applicable}
\item{AdRevenue}{Gross advertising revenue per advertising page in 2006 (in thousands of dollars)}
\item{Circulation}{Ciculation (in millions)}
}}
\usage{
data(AdRevenue)
}
\description{
U.S. Magazine Ad Revenue in 2006
}
\keyword{datasets}

