% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airfares}
\alias{airfares}
\title{One-way airfares}
\format{A data frame with 17 observations on the following 3 variables:
\describe{
\item{City}{Destination city ID}
\item{Fare}{One-way airfare (in U.S. dollars)}
\item{Distance}{Distance (in miles)}
}}
\usage{
data(airfares)
}
\description{
One-way airfares
}
\keyword{datasets}

