% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ais}
\alias{ais}
\title{Australian Institute of Sport}
\format{A data frame with 202 observations on the following 14 variables:
\describe{
\item{Sex}{0 = male, 1 = female}
\item{Ht}{Height in cm}
\item{Wt}{Weight in kg}
\item{RCC}{Red cell count}
\item{WCC}{White cell count}
\item{Hc}{Hematocrit}
\item{Hg}{Hemoglobin}
\item{Ferr}{Plasma ferritin concentration}
\item{BMI}{Body mass index = weight / height^2}
\item{SSF}{Sum of skin folds}
\item{Bfat}{Percent body fat}
\item{Label}{Label for sex and sport}
\item{Sport}{Sport}
}}
\usage{
data(ais)
}
\description{
Australian Institute of Sport
}
\keyword{datasets}

