% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banknote}
\alias{banknote}
\title{Swiss Bank Notes}
\format{A data frame with 200 observations on the following 7 variables:
\describe{
\item{Length}{length of the banknote (in mm)}
\item{Left}{length of the left edge of the banknote (in mm)}
\item{Right}{length of the right edge of the banknote (in mm)}
\item{Bottom}{distance from the image to the bottom edge (in mm)}
\item{Top}{distance from the image to the top edge (in mm)}
\item{Diagonal}{length of the diagonal (in mm)}
\item{Y}{Y=0 if counterfeit, and Y=1 if genuine}
}}
\usage{
data(banknote)
}
\description{
Swiss Bank Notes
}
\keyword{datasets}

