% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bayarea}
\alias{bayarea}
\title{Interest rates in the Bay Area}
\format{A data frame with 19 observations on the following 5 variables:
\describe{
\item{Month}{month index, 1-19, starting with June}
\item{MonthName}{name of month}
\item{InterestRate}{interest rate}
\item{LoansClosed}{the amount of loans closed (in millions of dollars)}
\item{VacancyIndex}{the vacancy index}
}}
\usage{
data(bayarea)
}
\description{
Interest rates in the Bay Area
}
\keyword{datasets}

