% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bimodal}
\alias{bimodal}
\title{50:50 mixture of Normals}
\format{A data frame with 20 observations on 1 variable:
\describe{
\item{x}{simulated values from a 50:50 mixture of N(-1, 1/9) and N(1, 1/9)}
}}
\usage{
data(bimodal)
}
\description{
50:50 mixture of Normals
}
\keyword{datasets}

