% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bonds}
\alias{bonds}
\title{US treasury bond prices}
\format{A data frame with 35 observations on the following 3 variables:
\describe{
\item{Case}{ID for U.S. Treasury bonds maturing between 1994 and 1998}
\item{CouponRate}{Coupon rate of the bond}
\item{BidPrice}{The bid prices are listed per "face value" of \$100 to be paid at maturity}
}}
\usage{
data(bonds)

data(bonds)
}
\description{
US treasury bond prices

US treasury bond prices
}
\keyword{datasets}

