% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bookstore}
\alias{bookstore}
\title{Monthly bookstore sales in Vienne, Austria}
\format{A data frame with 93 observations on the following 15 variables:
\describe{
\item{Sales}{Sales in hundreds of dollars}
\item{Advert}{Advertising spend in the current month}
\item{Lag1Advert}{Advertising spend in the previous month}
\item{Time}{Time in months}
\item{Month_2}{Dummy variable for month 2}
\item{Month_3}{Dummy variable for month 3}
\item{Month_4}{Dummy variable for month 4}
\item{Month_5}{Dummy variable for month 5}
\item{Month_6}{Dummy variable for month 6}
\item{Month_7}{Dummy variable for month 7}
\item{Month_8}{Dummy variable for month 8}
\item{Month_9}{Dummy variable for month 9}
\item{Month_10}{Dummy variable for month 10}
\item{Month_11}{Dummy variable for month 11}
\item{Month_12}{Dummy variable for month 12}
}}
\usage{
data(bookstore)
}
\description{
Monthly bookstore sales in Vienne, Austria
}
\keyword{datasets}

