% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bordeaux}
\alias{Bordeaux}
\title{Prices of Bordeaux wines}
\format{A data frame with 72 observations on the following 9 variables:
\describe{
\item{Wine}{Name of the winery (character vector)}
\item{Price}{price (in pounds sterling) of 12 bottles of wine}
\item{ParkerPoints}{Robert Parker’s rating (out of 100)}
\item{CoatesPoints}{Clive Coate’s rating (out of 20)}
\item{P95andAbove}{Is Parker’s score above 95? 1 = yes, 0 = no}
\item{FirstGrowth}{Is the wine a first growth?  1 = yes, 0 = no}
\item{CultWine}{Is the wine considered to be a cult favorite?  1 = yes, 0 = no}
\item{Pomerol}{Is the wine from Pomerol?  1 = yes, 0 = no}
\item{VintageSuperstar}{Is the wine a vintage superstar as awarded by Parker?  1 = yes, 0 = no}
}}
\usage{
data(Bordeaux)
}
\description{
Prices of Bordeaux wines
}
\keyword{datasets}

