% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modernregressiondata.R
\docType{package}
\name{resampledata}
\alias{resampledata}
\alias{resampledata-package}
\title{modernregressiondata: Data Sets for A Modern Approach to Regression with R}
\description{
Package of data sets from "A Modern Approach to Regression with R"
(2009) by Simon J. Sheather.
}
\section{Data sets}{

Type \code{data(package="modernregressiondata")} to view all datasets.
}
\examples{
library(modernregressiondata)
data(FlightDelays)
}
\references{
Sheather, Simon J. (2009). A Modern Approach
   to Regression with R. New York, NY: Springer.
}

